automate plot generation in main func

+++ plot.py
@@ -12,7 +12,7 @@ class PlotData(object):
         """
         self.plot_storage = os.path.join(imgdir_path, "img")
         if not os.path.exists(self.plot_storage):
-            os.mkdir(self.plot_storage)
+            os.makedirs(self.plot_storage)
 
     def plot(
         self,
@@ -22,6 +22,7 @@ class PlotData(object):
         plot: Callable,
         rotation: int = 0,
         figsize: Tuple[int]=(8, 6),
+        autosave: bool = False,
         **kwargs) -> None:
         """
         Function which takes data, x, y columns and sns or plt function and plot
@@ -37,12 +38,15 @@ class PlotData(object):
         plt.title(label=f"{x} vs {y} - {plot_name}", fontweight="bold")
         plt.xlabel(xlabel=x, fontweight="bold")
         plt.ylabel(ylabel=y, fontweight="bold")
-        plt.legend()
+        #plt.legend()
         save_path = os.path.join(self.plot_storage, f"{x}_{y}_{plot_name}.png")
         plt.savefig(save_path)
-        plt.show()
-        save_img = input("Do you want to save this plot? (y/n)")
-        # This has to be this way, because if plot is saved after plt.show()
-        # plain image is saved. plt.show() clears canvas after call
-        if save_img.lower() in ["n", "no", "nope"]:
-            os.remove(path=save_path)
+        if autosave == False:
+            plt.show()
+            save_img = input("Do you want to save this plot? (y/n)")
+            # This has to be this way, because if plot is saved after plt.show()
+            # plain image is saved. plt.show() clears canvas after call
+            if save_img.lower() in ["n", "no", "nope"]:
+                os.remove(path=save_path)
+        elif autosave == True:
+            print(f"Plot saved as {save_path}")
